+++ CHANGELOG.md
@@ -4,10 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [1.3.3] - 2022-08-22
 
 ### Fixed
 - checkout problem (in case of the same branch)
+- bugfix at github runner, relates to select right regex: [release](https://github.com/actions/runner/releases/tag/v2.295.0), [fix](https://github.com/actions/runner/pull/1794)
+
+### Added 
+- if regex sha is same to the current default head, take the last possible regex
 
 ## [1.3.2] - 2022-05-13